test.py: configurable authenticator and authorizer
For scylla servers, default to allow all for authenticator and
authorizer but allow to be configurable (e.g. PasswordAuthenticator and
CassandraAuthorizer).

Remove auth provider as this is not needed.

Disabling authentication avoids complications later for tests taking
down nodes as system_auth starts with RF=1 and it would require clusters
to increase RF and run repair.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>