test.py: harness for cluster management from test
Harness for managing cluster for Python tests.

Before this patch, the same cluster was used for all test cases in a
pytest file. This patch provides an API for tests to alter the topology
of the cluster. When this happens, the cluster is marked dirty. Both
cluster and driver connection are kept if the cluster is not dirty.

Fixtures pick up host and port from API instead of passed as params.

Fixtures were changed so the harness can do before and after test
checks and see if the cluster is dirty. The harness client helper is
passed a function so it can make driver connections on demand as it
makes new clusters. The cql fixture is removed as driver connection is
managed by the harness client. The scope of keyspace fixture keyspace is
now function as a new keyspace is needed when cluster changes.

For the ScyllaCluster, split the "cluster" list of active servers into
started (dict), stopped (dict), removed (list). The first two have host
id as keys, and removed is just a list of host ids.

Harness client helper manages all API requests to the server. For now,
a request to create a new cluster is just waited.

In the future, if a larger number of servers per cluster is required
and would timeout the HTTP API request, this could be changed to return
immediately and then the client checks until cluster is up
(_wait_for_cluster()).

Initial tests for topology and schema changes are added.

Fixture cql_test_connection is removed as it's now redundant.

API provides before/after test checks, test for harness and cluster up,
configured cql port, list of nodes, cycle current cluster, stop/start,
number of replicas (RF), manually mark cluster dirty. And node
operations: add, stop, start, restart, remove, decommission, replace.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>