gms, service: live endpoint copy method
Move replication logic for live endpoint across shards to a separate
method

This will be used by API get alive nodes.

As this is now in a method and outside gossiper::run(), assert it's
called from shard 0.

Also as this might now be called concurrently from API, protect the copy
with a rwlock.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>