test/pylib: ManagerClient helpers to wait for...
server to see other servers after start/restart

When starting/restarting a server, provide a way to wait for the server
to see at least n other servers.

Also leave the implementation method available for manual use and update
previous tests and a similar one to specify a specific server to be
seen.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>