test.py: allow pytest to run test case
Support running specific test cases within a pytest file.

The name pattern must match suite/test_file/test_case format. For
example:

    ./test.py -v --mode dev topology/test_random_tables/test_new_table

When there is a test_case component in one of the provided names, and if
any if the test_file itself was not matched, the code parses the file
(without importing) and finds the test cases, then tries to find
matches.

While there, this refactor fixes a bug of multiple name matches against
the same test file causing multiple test runs.

When a test is for a specific case in a test file, the unique name now
contains also the case in it (i.e. log file).

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>