test.py: allow pytest to run test case
Support running specific test cases within a pytest file.

The name pattern must match suite/test_file/test_case format. For
example:

    ./test.py -v --mode dev topology/test_random_tables/test_new_table

When there is a test_case component in one of the provided names, and if
any if the test_file itself was not matched, the code parses the file
(without importing) and finds the test cases, then tries to find
matches.

    # Single test for suite/test_file (which includes test_case)
    ./test.py -v --mode dev suite/test_file suite/test_file/test_case

While there, this refactor fixes a bug of multiple name matches against
the same test file causing multiple test runs. (missed break in the
for p in t loop):

    # Single test for suite/test_file (which includes test_case)
    ./test.py -v --mode dev suite/test_file suite/test_file

When a test is for a specific case in a test file, the unique name now
contains also the case in it (i.e. log file).

The test_case strings can be prefixes. For example:

    # Run all test cases starting with test_new in test_random_tables
    ./test.py -v --mode dev topology/test_random_tables/test_
    # Run all test cases in test_random_tables (separately/parallel)
    ./test.py -v --mode dev topology/test_random_tables/test_

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>