test.py: async workers
Before this patch, test.py would run one test (file) for each job. The
default number of jobs is the number of processors.

Test suites would define the amount of parallelism with the Pool size
and had no consideration of available resources.

On observation, the main limiting resource is not CPU but memory. In
particular, tests needing a Scylla server or cluster require significant
amounts of memory.

This patch changes the logic of running tests. It considers the amount
of memory available and prioritizes slow suites and tests accordingly.

There are many corner cases, but the 2 most common/relevant ones are:

1. just running a single suite and mode
2. running everything across all modes and suites

In the first case, we want to spread out the suite as much as possible
across the available hardware and run as many workers in parallel.

In the second case, in the beginning it's convenient to start early
multiple suites with run first tests, considering the mode and the suite
load.

Also note for the second case, at some point there will be only one
suite remaining, so the scenario becomes the second case and we want
this suite to be run across all the available hardware (more parallel
workers).

To achieve this, this patch introduces workers and the MemoryManager
class.

For now, workers are asyncio Tasks. But the plan is to make them
eventually multiprocessing Processes to avoid Python GIL contention.

Workers can run test for any suite. They pick suites to run from a
PriorityQueue, where the suites are ordered by a precedence list. Lower
values get picked first. The precedence list's first element is computed
with the mode order determined by mode_run_idx scaled by the suite load.
The second and third elements are the remaining run first tests and
total tests respectively, scaled down by the current amount of workers
actively processing the suite.

More than one worker can run tests for a suite. But once there are
already as many workers on a suite as there are test remaining, the
suite is not put back into the PriorityQueue.

Once the worker picks a suite it stays processing it until there are no
more tests remaining.

Suites can specify a minimum amount of memory to reserve. This should
satisfy the majority of tests in the suite. Outlier tests needing more
memory should be marked in suite.yaml in a section `test_extra_memory`
where more memory can be specified in MiB.

The values provided in this patch were picked after extensive testing
but they are not perfect. We should keep an eye for OOM or swapping for
a while. Or alternatively, to see if the values observed were too big.

The suite runner now manages the resources it needs to run a suite
(e.g. clusters), and disposes them once it's done. So there is no
longer need for the ArtifactRegistry.

The context for running a test are passed to the Test.run() as
a dict `test_args`. For example, `{"cluster": <ScyllaCluster>}`.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>