test.py: async suite runners
Before this patch, test.py would run one test (file) for each job. The
default number of jobs is the number of processors.

Test suites would define the amount of parallelism with the Pool size
and had no consideration of available resources.

On observation, the main limiting resource is not CPU but memory. In
particular, tests needing a Scylla server or cluster require significant
amounts of memory.

This patch changes the logic of running tests. It considers the amount
of memory available and prioritizes slow suites and tests accordingly.

There are many corner cases, but the 2 most relevant ones are either
running everything across all modes and suites, and just running a
single suite.

In the first case, it's convenient to run as many long running tests in
parallel as possible, across multiple suites. So it's convenient to run
as many tests of the same suite over the suite's resources (e.g. Scylla
cluster).

In the second case, we want to spread out the suite as much as possible
across the available hardware and run as many tests in parallel.

Also note for the first case, when there's only one suite with tests
remaining, it kind of becomes the second case.

To achieve this, this patch introduces workers, the MemoryManager class,
and suite runners.

For now, workers are asyncio Tasks. But the plan is to make them
multiprocessing Processes.

Workers can run test for any suite. They pick suites to run from a
PriorityQueue, where the suites are ordered by a precedence list. Lower
values get picked first. The precedence list's first element is the
mode index (i.e. debug<release<dev). The second element is the amount of
tests marked as run first for the suite (negative value for ordering).
The third element is all the remaining tests (also negative). So, for
example, suite A mode debug with 3 run first and 10 total tests will be
picked before suite B mode release with 4 run first and 20 total tests.

More than one worker can run tests for a suite. But once there are
already as many workers/runners on a suite, the suite is not put back
into the PriorityQueue.

To avoid having a single suite take all resources (first case above),
all the priority values are scaled times the amount of current workers
on that suite. So suites with no workers on it (or less) have precedence
over suites with workers.

Once the worker picks a suite it runs it. It reserves the memory for the
suite and before each test it reserves extra. In suite.yam there's now a
section test_memory where the default test memory requirement is
specified, and also the memory for test requiring more than that.

The values provided in this patch were picked after extensive testing
but they are not perfect. We should keep an eye for OOM or swapping for
a while. Or alternatively, to see if the values observed were too big.

The suite runner now manages the resources it needs (e.g. clusters).
There is no longer need for the ArtifactRegistry.

The context for running a test are passed to the Test.run() as
test_args. For example, {"cluster": ScyllaCluster}.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>