test.py: Change to test naming to suite::test::case
Change test naming to suite::test::case to allow filtering by specific
case name.

Naming a test in command line should either include the suite or have
the '::<test name or substring>' format.

Test and case names can be just a substring of the wanted test(s).

    boost::database_test::clear_snapshot       runs one test case
    topology::test_change_ip                   runs all tests in a file
    cql-pytest                                 suite only, runs all its tests
    ::test_tablets                             tests matching any suite

Matching is done specifically for each of the 3 name components.
Unspecified either of suite, test, or case, act as a wildcard.

Representation omits the '::case' part if the case is not specified. For a
wildcard pattern, if no suite is specified a prefix 'all::' is used.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>